이동에 대한 메소드 클래스로 분리
