+++ src/main/java/mission1/racing/RacingGame.java
@@ -5,31 +5,15 @@ import java.util.Scanner;
 public class RacingGame {
     private int time;
     private int carNumber;
-    private int[] carPositions = {0, 0, 0};
-    
-    RacingRandom random = new RacingRandom();
+
+    Movement movement = new Movement();
 
     public void startGame() {
         setGame();
 
         for(int i = 0; i < time; i++) {
-            move();
+            movement.move();
         }
-    }
-
-    public void move() {
-        for(int i = 0; i < carPositions.length; i++) {
-            carPositions[i] += moveCondition();
-            showGameStatus(i);
-        }
-    }
-
-    private int moveCondition() {
-        if (random.random() >= 4) {
-            return 1;
-        }
-
-        return 0;
     }
 
     private void setGame() {
@@ -40,12 +24,5 @@ public class RacingGame {
 
         System.out.println("시도할 회수는 몇 회 인가요?");
         time = scanner.nextInt();
-    }
-
-    private void showGameStatus(int i) {
-        for(int j = 0; j < carPositions[i]; j++) {
-            System.out.print("-");
-        }
-        System.out.print("\n");
     }
 }
